StepTalk scripts panel: int->NSInteger transition
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/steptalk/trunk@36661 72102866-910b-0410-8b05-ffd578937521

+++ ApplicationScripting/Source/STScriptsPanel.m
@@ -188,15 +188,15 @@ STScriptsPanel *sharedScriptsPanel = nil;
     }
 }
 
-- (int)       browser: (NSBrowser *) sender
- numberOfRowsInColumn: (int) column
+- (NSInteger) browser: (NSBrowser *) sender
+ numberOfRowsInColumn: (NSInteger) column
 {
     return [scripts count];
 }
 - (void) browser: (NSBrowser *) sender
  willDisplayCell: (NSBrowserCell *) cell
-           atRow: (int) row
-          column: (int) column
+           atRow: (NSInteger) row
+          column: (NSInteger) column
 {
     NSString *name;
     if(sender != scriptList)
